pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vision-coder-openenv"
version = "0.1.0"
description = "Screenshot-to-HTML reinforcement learning environment (OpenEnv-compatible)"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "httpx>=0.27.0",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
    "datasets>=2.18.0",
    "beautifulsoup4>=4.12.0",
    "transformers>=4.40.0",
    "torch>=2.2.0",
    "openai>=1.20.0",
    "playwright>=1.43.0",
]

[tool.setuptools]
packages = ["openenv", "openenv.server", "vcoder", "vcoder.data", "vcoder.rewards"]

[tool.setuptools.package-dir]
"openenv" = "."
"openenv.server" = "server"
"vcoder" = "vcoder"
"vcoder.data" = "vcoder/data"
"vcoder.rewards" = "vcoder/rewards"
